fix: bringing specific metadata to these pages
