+++ public/sitemap-0.xml
@@ -1,14 +1,14 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:news="http://www.google.com/schemas/sitemap-news/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml" xmlns:mobile="http://www.google.com/schemas/sitemap-mobile/1.0" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" xmlns:video="http://www.google.com/schemas/sitemap-video/1.1">
-<url><loc>https://gap.karmahq.xyz/sitemaps/projects/sitemap.xml</loc><lastmod>2024-08-28T19:00:03.175Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://gap.karmahq.xyz/api/sponsored-txn</loc><lastmod>2024-08-28T19:00:03.175Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://gap.karmahq.xyz/sitemaps/communities/sitemap.xml</loc><lastmod>2024-08-28T19:00:03.175Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://gap.karmahq.xyz/my-projects</loc><lastmod>2024-08-28T19:00:03.175Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://gap.karmahq.xyz/admin/communities</loc><lastmod>2024-08-28T19:00:03.175Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://gap.karmahq.xyz/funding-map/add-program</loc><lastmod>2024-08-28T19:00:03.175Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://gap.karmahq.xyz/admin</loc><lastmod>2024-08-28T19:00:03.175Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://gap.karmahq.xyz/admin/communities/stats</loc><lastmod>2024-08-28T19:00:03.175Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://gap.karmahq.xyz/funding-map/manage-programs</loc><lastmod>2024-08-28T19:00:03.175Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://gap.karmahq.xyz</loc><lastmod>2024-08-28T19:00:03.175Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://gap.karmahq.xyz/funding-map</loc><lastmod>2024-08-28T19:00:03.175Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://gap.karmahq.xyz/sitemaps/communities/sitemap.xml</loc><lastmod>2024-08-28T19:27:44.160Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://gap.karmahq.xyz/admin</loc><lastmod>2024-08-28T19:27:44.161Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://gap.karmahq.xyz/api/sponsored-txn</loc><lastmod>2024-08-28T19:27:44.161Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://gap.karmahq.xyz/sitemaps/projects/sitemap.xml</loc><lastmod>2024-08-28T19:27:44.161Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://gap.karmahq.xyz/funding-map/add-program</loc><lastmod>2024-08-28T19:27:44.161Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://gap.karmahq.xyz/my-projects</loc><lastmod>2024-08-28T19:27:44.161Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://gap.karmahq.xyz/admin/communities</loc><lastmod>2024-08-28T19:27:44.161Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://gap.karmahq.xyz/admin/communities/stats</loc><lastmod>2024-08-28T19:27:44.161Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://gap.karmahq.xyz</loc><lastmod>2024-08-28T19:27:44.161Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://gap.karmahq.xyz/funding-map</loc><lastmod>2024-08-28T19:27:44.161Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://gap.karmahq.xyz/funding-map/manage-programs</loc><lastmod>2024-08-28T19:27:44.161Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
 </urlset>
